feat(auto-trigger): T36-T39 · 撤回窗口浮动通知卡 (体验要素 #30) · v0.4.0 收官
最后一块拼图. 用户能完整端到端看到"触发 → 5min 撤回 → 真生成 / 取消"流程.
不靠 SSE, 用 3s polling 兜底, 体验已可用 (SSE 留 v0.4.1).

新增端点 (后端):
- POST /projects/:pid/auto-triggers/rules/:rid/test-fire (T36)
  立即创建 scheduled execution, fire_at = now+30s (demo 用短窗口);
  真实场景 cron 5min, 太长不好演示
- svc.schedule_execution: withdraw_window_min int → float (支持 0.5 等小数)

新增前端组件:
- src/features/auto-triggers/WithdrawNotification.tsx (T37)
  * AppLayout 顶层全局挂载, 右下角浮动
  * 3s polling 查 status=scheduled 的执行
  * 倒计时每秒更新; <30s 变橙, <10s 变红, 归零变绿"已生成"
  * 取消按钮调 cancel API, optimistic UI 变红 "已取消" 1.5s 后消失
  * 支持多条聚合 (按 fire_at 排列)

- AutoTriggersPanel.tsx 抽屉详情新增 (T38):
  * 演示提示卡 + "立即触发演示" 按钮
  * 仅 rule.enabled=true 时显示
  * 触发后 invalidate scheduled query → 浮动卡 3s 内自动出现

端到端验证 (浏览器实测):
1. 点立即触发演示 → 30s 倒计时启动
2. 关闭抽屉看右下角浮动通知
3. 不取消的话: withdraw_resolver 真 fire → 真调 LLM → asset 入库
   + 通知卡显示 "已生成 · 撤回窗口已过 · LLM 调用进行中" (绿色)
4. 取消的话: 通知卡变红 "已取消 · LLM 已被阻止调用 · 无 token 消耗" 1.5s 后消失

测试:
- test_auto_trigger_http.py 新增 4 个 test-fire cases
  (happy + 404 missing + 404 cross-project + 实例级 fan-out 到任意项目)
- 全 backend 960/960; 全前端 711/711

v0.4.0 完整 task 状态 (10 个全部 ✅):
- T26 DB Schema (47)
- T27 APScheduler (13)
- T28 默认规则 seeder (14)
- T29 cron RuleEvaluator (24)
- T30+T31 dispatcher + withdraw_resolver (11)
- T32 REST API (26)
- T33+T34+T35 前端 UI (711 前端 tests)
- T36+T37+T38 撤回窗口通知卡 (4 tests · 体验要素 #30) ← 本次

v0.4.0 共 12 个 commit; ~140 后端新测试 / 711 前端全绿.

+++ code/tokenknows-api/app/services/auto_trigger_service.py
@@ -248,7 +248,7 @@ def schedule_execution(
     project_id: str,
     signal: TriggerSignal,
     *,
-    withdraw_window_min: int = DEFAULT_WITHDRAW_WINDOW_MIN,
+    withdraw_window_min: float = DEFAULT_WITHDRAW_WINDOW_MIN,
     evaluation: Any = None,
 ) -> TriggerExecution:
     """规则命中后调用, 写一条 status=scheduled 的执行记录.
